+++ NEWS.md
@@ -1,3 +1,16 @@
+# Package update 0.1.4 -> 0.1.5
+
+## New functions: 
+* New functions for price imputations (*data_imputing*) and for detecting new and disappearing products (*products* and *products_fig*).
+
+## Improvements: 
+* The **data_preparing** function has now two additional parameters which allow to consider zero prices and zero quantities.
+* New methods for **elasticity** and **elasticity_fig** functions have been added.
+
+## Bug fixes: 
+* a new parameter has beed added to the function 'price_indices' (*names*). It allows to compare e.g. the same splice multilateral index but for many different splicing methods (their names can be distinguished).
+* The *geksj* function has been corrected for the case when we observe large number of decreasing prices.
+
 # Package update 0.1.3 -> 0.1.4
 
 ## New functions